Move label and description to rdfWrapper too.
